requirements.txt
# Core ML / embeddings
sentence-transformers==3.3.1

# Vector store
chromadb==0.6.3

# Google Gemini
google-genai==1.16.0

# Data
numpy==2.1.3
pandas==2.2.3

# HTTP / caching
requests==2.32.3

# Evaluation
rouge-score==0.1.2
scikit-learn==1.5.2

# Jupyter / visualization
jupyterlab==4.4.3
ipykernel==6.29.5
matplotlib==3.9.3
seaborn==0.13.2

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
